backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Harsht139", email = "harsh.t@dataslush.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "google-cloud-aiplatform>=1.138.0",
    "google-genai>=0.1.0",
    "fpdf2>=2.8.2",
    "httpx>=0.28.1",
    "llama-index>=0.14.15",
    "llama-index-core>=0.14.15",
    "llama-index-embeddings-google-genai>=0.3.0",
    "llama-index-llms-google-genai>=0.3.0",
    "llama-index-readers-file>=0.5.6",
    "mypy>=1.19.1",
    "pandas>=2.0.0",
    "pydantic>=2.12.5",
    "pypdf>=4.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "ruff>=0.15.1",
    "tenacity>=8.0.0",
    "uvicorn>=0.41.0",
    "google-api-python-client>=2.116.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "markdown>=3.10.2",
    "supabase>=2.10.0",
    "docling>=2.74.0",
    "chromadb>=1.5.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
